Rename setup validator and organize tests
