+++ AGENTS.md
@@ -104,7 +104,7 @@ See `progress/README.md`. Default: no progress file. Create only for hardware-fa
 Smallest meaningful gate first, then broaden:
 
 ```bash
-python -m pytest tests/test_deck_loader.py tests/test_holder_labware.py -q
+python -m pytest tests/deck/test_deck_loader.py tests/deck/test_holder_labware.py -q
 python -m pytest tests/protocol_engine -q
 python -m pytest -q
 python setup/validate_setup.py configs/gantry/cub_xl_asmi.yaml configs/deck/asmi_deck.yaml configs/protocol/asmi/indentation.yaml
