+++ docs/reference/src/protocol_engine/setup_validator.md
@@ -1,0 +1,5 @@
+# `src.protocol_engine.setup_validator`
+
+<!-- Auto-generated by docs/gen_ref_pages.py. -->
+
+::: src.protocol_engine.setup_validator
