+++ tests/deck/__init__.py
@@ -1,0 +1,1 @@
+"""Deck and labware tests."""
