+++ tests/instruments/uvvis_ccs/__init__.py
@@ -1,0 +1,1 @@
+"""UV/VIS CCS instrument tests."""
